fix: login, bcrypt function comparePassword, and refactor login, change route from users/auth/login to auth/login
