+++ controllers/users.js
@@ -1,61 +1,6 @@
-
-
 /* eslint-disable no-unused-vars */
-const { body, validationResult } = require('express-validator');
-const { allUsers, deleteUser, findUsers } = require('../services/user');
-const { comparePassword } = require('../helpers/bcrypt');
 
-const login = async (req, res) => {
-  const { email, password } = req.body;
-
-  body(email).isEmail(),
-
-  body(password)
-    .isLength({ min: 8 })
-    .matches(/\d/)
-    .matches('[A-Z]')
-    .trim()
-    .trim();
-
-  const errors = validationResult(req);
-
-  if (!errors.isEmpty()) {
-    res.status(400).json({ errors: errors.array() });
-  }
-
-  try {
-    const userFound = await findUsers(email);
-
-    if (!userFound) {
-      res.status(404).send('Email not registered');
-    }
-
-    await comparePassword(password, userFound.password)
-      .then((result) => {
-        if (result) {
-          res.status(201).json({
-            msg: 'Login successfully',
-            user: userFound,
-          });
-        } else {
-          res.status(404).json({
-            msg: 'The password is wrong',
-          });
-        }
-      })
-      .catch((error) => {
-        res.status(404).json({
-          ok: false,
-          error,
-        });
-      });
-  } catch (error) {
-    res.status(404).json({
-      ok: false,
-      error,
-    });
-  }
-};
+const { allUsers, deleteUser } = require('../services/user');
 
 
 const getAllUsers = async (req, res) => {
@@ -87,7 +32,4 @@ const deleteSingleUser = async (_req, res, next) => {
   });
 };
 
-module.exports = {
-  login,
-  deleteSingleUser,
-};
+module.exports = { getAllUsers, deleteSingleUser };
