+++ routes/auth.js
@@ -3,8 +3,9 @@ const express = require('express');
 const router = express.Router();
 
 const { validateUser } = require('../middlewares/validators');
-const { createUser } = require('../controllers/auth');
+const { createUser, login } = require('../controllers/auth');
+
 
 router.post('/register', validateUser, createUser);
-
+router.post('/login', login);
 module.exports = router;
